# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_asset_api_v1_assets_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "JPG、JPEG 或 PNG 图片",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_asset_api_v1_assets_post",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SourceAssetResponse": {
        "properties": {
          "assetId": {
            "title": "Assetid",
            "type": "string"
          },
          "contentUrl": {
            "title": "Contenturl",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "duplicateOfAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicateofassetid"
          },
          "fileHash": {
            "title": "Filehash",
            "type": "string"
          },
          "fileName": {
            "title": "Filename",
            "type": "string"
          },
          "fileSize": {
            "title": "Filesize",
            "type": "integer"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "mediaType": {
            "title": "Mediatype",
            "type": "string"
          },
          "storageKey": {
            "title": "Storagekey",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "assetId",
          "fileName",
          "mediaType",
          "storageKey",
          "fileHash",
          "width",
          "height",
          "fileSize",
          "contentUrl",
          "createdAt"
        ],
        "title": "SourceAssetResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "这是提供给 AI 与开发者的配置接口，不是教师日常操作页面。当前支持：上传单页图片、教师框题、OCR、人工修订、来源追溯和发布飞书 Base。",
    "title": "小学数学错题学习系统 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets": {
      "post": {
        "operationId": "upload_asset_api_v1_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_asset_api_v1_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceAssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "上传图片超过系统允许的大小。"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "上传文件不是支持的图片格式。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "上传一张作业或试卷图片",
        "tags": [
          "来源图片"
        ]
      }
    },
    "/api/v1/assets/{asset_id}": {
      "get": {
        "operationId": "get_asset_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceAssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "读取来源图片元数据",
        "tags": [
          "来源图片"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/content": {
      "get": {
        "operationId": "get_asset_content_api_v1_assets__asset_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "读取来源图片内容",
        "tags": [
          "来源图片"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/detection-runs": {
      "post": {
        "operationId": "create_detection_run_api_v1_assets__asset_id__detection_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDetectionRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "外部识别服务返回了无法解析的结果。"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "识别或发布服务暂时不可用，或尚未正确配置。"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "外部识别服务运行超时。"
          }
        },
        "summary": "可选：自动检测题目候选框",
        "tags": [
          "题目录入"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/problems": {
      "get": {
        "operationId": "list_asset_problems_api_v1_assets__asset_id__problems_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetProblemCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "读取这张来源页中的全部错题",
        "tags": [
          "题目审核"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/regions": {
      "post": {
        "operationId": "create_region_api_v1_assets__asset_id__regions_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "保存一个人工题目框",
        "tags": [
          "题目录入"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/regions/batch": {
      "post": {
        "operationId": "create_regions_batch_api_v1_assets__asset_id__regions_batch_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRegionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRegionCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "当前状态不允许执行这个操作。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "批量保存教师确认的题目框",
        "tags": [
          "题目录入"
        ]
      }
    },
...
    "/api/v1/problems/{problem_id}": {
      "get": {
        "operationId": "get_problem_api_v1_problems__problem_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "problem_id",
            "required": true,
            "schema": {
              "title": "Problem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizedProblemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "读取一条完整的规范化题目记录",
        "tags": [
          "题目记录"
        ]
      }
    },
    "/api/v1/problems/{problem_id}/publications/lark": {
      "post": {
        "operationId": "publish_problem_api_v1_problems__problem_id__publications_lark_post",
        "parameters": [
          {
            "in": "path",
            "name": "problem_id",
            "required": true,
            "schema": {
              "title": "Problem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemPublicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "当前状态不允许执行这个操作。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "外部识别服务返回了无法解析的结果。"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "识别或发布服务暂时不可用，或尚未正确配置。"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "外部识别服务运行超时。"
          }
        },
        "summary": "把已整理题目发布到飞书 Base",
        "tags": [
          "飞书 Base"
        ]
      }
    },
    "/api/v1/regions/{region_id}/crop": {
      "get": {
        "operationId": "get_region_crop_api_v1_regions__region_id__crop_get",
        "parameters": [
          {
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          }
        },
        "summary": "读取题目裁图",
        "tags": [
          "题目录入"
        ]
      }
    },
    "/api/v1/regions/{region_id}/ocr-runs": {
      "post": {
        "operationId": "create_ocr_run_api_v1_regions__region_id__ocr_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "HTTP 请求或上传表单格式错误。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "找不到请求的来源材料或题目记录。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求内容没有通过校验。"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "系统内部或本地文件存储发生错误。"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "外部识别服务返回了无法解析的结果。"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "识别或发布服务暂时不可用，或尚未正确配置。"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "外部识别服务运行超时。"
          }
        },
        "summary": "新增一次 OCR 识别",
        "tags": [
          "OCR 与修订"
        ]
      }
    },
...
  },
...
}